state: storage: task-assignment: Refactor over rkyv
