+++ crates/relayer-types/types-gossip/src/lib.rs
@@ -8,373 +8,20 @@
 #![deny(clippy::needless_pass_by_ref_mut)]
 #![deny(clippy::missing_docs_in_private_items)]
 
-use std::time::{SystemTime, UNIX_EPOCH};
+mod cluster;
+mod peer_id;
+mod peer_info;
 
-use base64::{Engine as _, engine::general_purpose::STANDARD as b64_standard};
-use derivative::Derivative;
-use ed25519_dalek::{
-    Digest, Keypair, PublicKey, SECRET_KEY_LENGTH, SecretKey, Sha512, Signature, SignatureError,
-};
-use libp2p::{Multiaddr, PeerId};
-use libp2p_identity::{
-    ParseError as PeerIdParseError, ed25519::Keypair as LibP2PKeypair,
-    ed25519::SecretKey as LibP2PSecretKey,
-};
-use rand_core::OsRng;
-use serde::{
-    Deserialize, Serialize,
-    de::{Error as SerdeErr, Visitor},
-};
-use std::{
-    fmt::{Display, Formatter, Result as FmtResult},
-    ops::Deref,
-    str::FromStr,
-};
-use util::{networking::is_dialable_multiaddr, raw_err_str};
-
-/// Returns the current unix timestamp in milliseconds
-fn get_current_time_millis() -> u64 {
-    SystemTime::now().duration_since(UNIX_EPOCH).expect("negative timestamp").as_millis() as u64
-}
-
-// -------------
-// | Peer Info |
-// -------------
-
-/// The topic prefix for the cluster management pubsub topic
-///
-/// The actual topic name will have the cluster ID postfixed; i.e.
-///     cluster-management-{cluster_id}
-pub const CLUSTER_MANAGEMENT_TOPIC_PREFIX: &str = "cluster-management";
-
-/// Contains information about connected peers
-#[derive(Clone, Debug, Serialize, Deserialize, Derivative)]
-#[derivative(PartialEq, Eq)]
-pub struct PeerInfo {
-    /// The identifier used by libp2p for a peer
-    pub peer_id: WrappedPeerId,
-    /// The multiaddr of the peer
-    pub addr: Multiaddr,
-    /// Last time a successful heartbeat was received from this peer
-    pub last_heartbeat: u64,
-    /// The ID of the cluster the peer belongs to
-    pub cluster_id: ClusterId,
-    /// The signature of the peer's ID with their cluster private key, used to
-    /// prove that the peer is a valid cluster member
-    #[derivative(PartialEq = "ignore")]
-    pub cluster_auth_signature: Vec<u8>,
-}
-
-impl Default for PeerInfo {
-    fn default() -> Self {
-        Self {
-            peer_id: WrappedPeerId(PeerId::random()),
-            addr: Multiaddr::empty(),
-            last_heartbeat: 0,
-            cluster_id: ClusterId("0".to_string()),
-            cluster_auth_signature: vec![],
-        }
-    }
-}
-
-impl PeerInfo {
-    /// Construct a new PeerInfo object
-    pub fn new(
-        peer_id: WrappedPeerId,
-        cluster_id: ClusterId,
-        addr: Multiaddr,
-        cluster_auth_signature: Vec<u8>,
-    ) -> Self {
-        Self {
-            addr,
-            peer_id,
-            cluster_id,
-            cluster_auth_signature,
-            last_heartbeat: get_current_time_millis(),
-        }
-    }
-
-    /// Construct a new PeerInfo object using the cluster private key
-    pub fn new_with_cluster_secret_key(
-        peer_id: WrappedPeerId,
-        cluster_id: ClusterId,
-        addr: Multiaddr,
-        cluster_keypair: &Keypair,
-    ) -> Self {
-        // Generate an auth signature for the cluster
-        let mut hash_digest = Sha512::new();
-        hash_digest.update(serde_json::to_vec(&peer_id).unwrap());
-        let sig = cluster_keypair.sign_prehashed(hash_digest, None /* context */).unwrap();
-
-        Self::new(peer_id, cluster_id, addr, sig.to_bytes().to_vec())
-    }
-
-    /// Verify that the signature on the peer's info is correct
-    pub fn verify_cluster_auth_sig(&self) -> Result<(), SignatureError> {
-        let sig = Signature::from_bytes(&self.cluster_auth_signature)
-            .map_err(|_| SignatureError::new())?;
-        let pubkey = self.cluster_id.get_public_key().map_err(|_| SignatureError::new())?;
-
-        // Hash the peer ID and verify the signature
-        let mut hash_digest = Sha512::new();
-        hash_digest.update(serde_json::to_vec(&self.peer_id).unwrap());
-        pubkey.verify_prehashed(hash_digest, None, &sig)
-    }
-
-    /// Getters and Setters
-    pub fn get_peer_id(&self) -> WrappedPeerId {
-        self.peer_id
-    }
-
-    /// Get the address stored in the PeerInfo
-    pub fn get_addr(&self) -> Multiaddr {
-        self.addr.clone()
-    }
-
-    /// Returns whether or not the peer's address is dialable
-    pub fn is_dialable(&self, allow_local: bool) -> bool {
-        is_dialable_multiaddr(&self.addr, allow_local)
-    }
-
-    /// Get the ID of the cluster this peer belongs to
-    pub fn get_cluster_id(&self) -> ClusterId {
-        self.cluster_id.clone()
-    }
-
-    /// Records a successful heartbeat
-    pub fn successful_heartbeat(&mut self) {
-        self.last_heartbeat = get_current_time_millis();
-    }
-
-    /// Get the last time a heartbeat was recorded for this peer
-    pub fn get_last_heartbeat(&self) -> u64 {
-        self.last_heartbeat
-    }
-}
-
-/// Wraps PeerID so that we can implement various traits on the type
-#[derive(Clone, Copy, Debug, Hash, Eq, PartialEq, PartialOrd, Ord)]
-pub struct WrappedPeerId(pub PeerId);
-
-impl WrappedPeerId {
-    /// Create a random PeerID and wrap it
-    pub fn random() -> Self {
-        Self(PeerId::random())
-    }
-
-    /// Get the underlying peer ID
-    pub fn inner(&self) -> PeerId {
-        self.0
-    }
-}
-
-impl Default for WrappedPeerId {
-    fn default() -> Self {
-        let skey = LibP2PSecretKey::try_from_bytes(&mut vec![0; SECRET_KEY_LENGTH]).unwrap();
-        let keypair = LibP2PKeypair::from(skey);
-        let peer_id = PeerId::from_public_key(&keypair.public().into());
-        Self(peer_id)
-    }
-}
-
-impl FromStr for WrappedPeerId {
-    type Err = PeerIdParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(WrappedPeerId(PeerId::from_str(s)?))
-    }
-}
-
-/// Deref so that the wrapped type can be referenced
-impl Deref for WrappedPeerId {
-    type Target = PeerId;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Display for WrappedPeerId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> FmtResult {
-        self.0.fmt(f)
-    }
-}
-
-/// Serialize PeerIDs
-impl Serialize for WrappedPeerId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let bytes = self.to_bytes();
-        serializer.serialize_bytes(&bytes)
-    }
-}
-
-/// Deserialize PeerIDs
-impl<'de> Deserialize<'de> for WrappedPeerId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        deserializer.deserialize_seq(PeerIDVisitor)
-    }
-}
-
-/// Visitor struct for help deserializing PeerIDs
-struct PeerIDVisitor;
-impl<'de> Visitor<'de> for PeerIDVisitor {
-    type Value = WrappedPeerId;
-
-    fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
-        formatter.write_str("a libp2p::PeerID encoded as a byte array")
-    }
-
-    fn visit_bytes<E>(self, v: &[u8]) -> Result<Self::Value, E>
-    where
-        E: SerdeErr,
-    {
-        PeerId::from_bytes(v)
-            .map(WrappedPeerId)
-            .map_err(|e| SerdeErr::custom(format!("deserializing byte array to PeerID: {e:?}")))
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut bytes_vec = Vec::new();
-        while let Some(value) = seq.next_element()? {
-            bytes_vec.push(value);
-        }
-
-        if let Ok(peer_id) = PeerId::from_bytes(&bytes_vec[..]) {
-            return Ok(WrappedPeerId(peer_id));
-        }
-
-        Err(SerdeErr::custom("deserializing byte array to PeerID"))
-    }
-}
-
-// ----------------
-// | Cluster Info |
-// ----------------
-
-/// A type alias for the cluster identifier
-#[derive(Clone, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ClusterId(String);
-
-impl ClusterId {
-    /// Construct a clusterID, it's more readable and debuggable to compress the
-    /// public key into a base64 encoded representation than to use the value
-    /// directly
-    pub fn new(cluster_pubkey: &PublicKey) -> Self {
-        #[allow(deprecated)]
-        let encoded_key = base64::encode(cluster_pubkey.as_bytes());
-        Self(encoded_key)
-    }
-
-    /// Get the cluster management pubsub topic name for the cluster identified
-    pub fn get_management_topic(&self) -> String {
-        format!("{CLUSTER_MANAGEMENT_TOPIC_PREFIX}-{}", self.0)
-    }
-
-    /// Get the public key represented by this cluster
-    pub fn get_public_key(&self) -> Result<PublicKey, SignatureError> {
-        #[allow(deprecated)]
-        let decoded_key = base64::decode(&self.0).map_err(|_| SignatureError::new())?;
-        PublicKey::from_bytes(&decoded_key)
-    }
-}
-
-impl Display for ClusterId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl FromStr for ClusterId {
-    // Conversion does not error
-    type Err = ();
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self(s.to_string()))
-    }
-}
-
-/// A wrapped cluster keypair which allows us to implement traits on the type
-#[derive(Debug)]
-pub struct ClusterAsymmetricKeypair(pub Keypair);
-
-impl ClusterAsymmetricKeypair {
-    /// Construct a new ClusterAsymmetricKeypair
-    pub fn new(keypair: Keypair) -> Self {
-        Self(keypair)
-    }
-
-    /// Get a cluster key from a base64 encoded string representing the secret
-    /// key
-    pub fn from_base64(s: &str) -> Result<Self, String> {
-        // Decode the base64 string and pad it to the correct length
-        let decoded =
-            b64_standard.decode(s).map_err(raw_err_str!("error decoding cluster key: {}"))?;
-
-        // Decompress the bytes into an Ed25519 keypair
-        let secret = SecretKey::from_bytes(&decoded)
-            .map_err(raw_err_str!("error decompressing cluster key: {}"))?;
-        let public = PublicKey::from(&secret);
-        let keypair = Keypair { secret, public };
-        Ok(Self(keypair))
-    }
-
-    /// Generate a random cluster keypair
-    pub fn random() -> Self {
-        let mut rng = OsRng {};
-        Self(Keypair::generate(&mut rng))
-    }
-}
-
-impl Clone for ClusterAsymmetricKeypair {
-    fn clone(&self) -> Self {
-        let cloned = Keypair::from_bytes(&self.0.to_bytes()).unwrap();
-        Self(cloned)
-    }
-}
-
-impl Deref for ClusterAsymmetricKeypair {
-    type Target = Keypair;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-// -----------
-// | Helpers |
-// -----------
-
-/// Mocks for peer info types
 #[cfg(feature = "mocks")]
-pub mod mocks {
-    use std::{
-        net::{IpAddr, Ipv4Addr},
-        str::FromStr,
-    };
+pub mod mocks;
 
-    use libp2p::Multiaddr;
+// Re-exports
+pub use cluster::{CLUSTER_MANAGEMENT_TOPIC_PREFIX, ClusterAsymmetricKeypair, ClusterId};
+pub use peer_id::WrappedPeerId;
+pub use peer_info::PeerInfo;
 
-    use super::{ClusterId, PeerInfo, WrappedPeerId};
-
-    /// Build a mock peer's info
-    pub fn mock_peer() -> PeerInfo {
-        // Build an RPC message to add a peer
-        let cluster_id = ClusterId::from_str("1234").unwrap();
-        let peer_id = WrappedPeerId::random();
-        let addr = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
-        let addr = Multiaddr::from(addr);
-
-        PeerInfo::new(peer_id, cluster_id, addr.clone(), vec![] /* signature */)
-    }
-}
+#[cfg(feature = "rkyv")]
+pub use peer_id::ArchivedWrappedPeerId;
 
 #[cfg(test)]
 mod tests {
