+++ crates/state/src/storage/tx/task_assignments.rs
@@ -1,47 +1,97 @@
 //! Handles storage operations on the table that tracks the assignment of tasks
 //! to nodes
+//!
+//! Task assignments use a per-task key scheme for efficient O(1) writes:
+//! - `assigned-task/{peer_id}/{task_id}` -> `()` for peer-to-task index
+//! - `task-assignment/{task_id}` -> `peer_id` for task-to-peer lookup
 
 // -----------
 // | Getters |
 // -----------
 
+use libmdbx::{RW, TransactionKind};
 use types_gossip::WrappedPeerId;
 use types_tasks::TaskIdentifier;
-use libmdbx::{RW, TransactionKind};
 
-use crate::{TASK_ASSIGNMENT_TABLE, storage::error::StorageError};
+use crate::{
+    TASK_ASSIGNMENT_TABLE,
+    storage::{ArchivedValue, error::StorageError},
+};
 
 use super::StateTxn;
 
-/// Get the key for the list of tasks assigned to a node
-fn node_tasks_key(id: &WrappedPeerId) -> String {
-    format!("assigned-tasks-{id}")
+/// A type alias for an archived peer ID value
+pub type PeerIdValue<'a> = ArchivedValue<'a, WrappedPeerId>;
+
+// ------------------
+// | Key Formatting |
+// ------------------
+
+/// The prefix for the peer-to-task index keys
+const ASSIGNED_TASK_PREFIX: &str = "assigned-task/";
+
+/// Get the key for a single task assignment (peer -> task mapping)
+fn assigned_task_key(peer_id: &WrappedPeerId, task_id: &TaskIdentifier) -> String {
+    format!("{ASSIGNED_TASK_PREFIX}{peer_id}/{task_id}")
 }
 
-/// Get the key for a task's assignment
-fn task_assignment_key(task_id: &TaskIdentifier) -> String {
-    format!("task-assignment-{task_id}")
+/// Get the prefix for all tasks assigned to a peer
+fn assigned_task_prefix(peer_id: &WrappedPeerId) -> String {
+    format!("{ASSIGNED_TASK_PREFIX}{peer_id}/")
 }
+
+/// Get the key for a task's assignment (task -> peer lookup)
+fn task_assignment_key(task_id: &TaskIdentifier) -> String {
+    format!("task-assignment/{task_id}")
+}
+
+/// Parse a task identifier from an assigned task key
+///
+/// Key format: `assigned-task/{peer_id}/{task_id}`
+fn parse_task_id_from_key(key: &str) -> Option<TaskIdentifier> {
+    // Find the last '/' and parse everything after it as a TaskIdentifier
+    let suffix = key.rsplit('/').next()?;
+    suffix.parse().ok()
+}
+
+// -----------
+// | Getters |
+// -----------
 
 impl<T: TransactionKind> StateTxn<'_, T> {
     /// Get the tasks assigned to a node
+    ///
+    /// This uses a cursor scan over the `assigned-task/{peer_id}/` prefix.
+    /// This operation is O(n) in the number of tasks assigned to the peer,
+    /// but is only used during peer failure recovery (rare).
     pub fn get_assigned_tasks(
         &self,
         peer_id: &WrappedPeerId,
     ) -> Result<Vec<TaskIdentifier>, StorageError> {
-        let key = node_tasks_key(peer_id);
-        self.read_set(TASK_ASSIGNMENT_TABLE, &key)
+        let prefix = assigned_task_prefix(peer_id);
+        let cursor = self
+            .inner()
+            .cursor::<String, ()>(TASK_ASSIGNMENT_TABLE)?
+            .with_key_filter(|key| key.starts_with(prefix.as_str()));
+
+        let mut tasks = Vec::new();
+        for entry in cursor.into_iter() {
+            let (key, _) = entry?;
+            if let Some(task_id) = parse_task_id_from_key(&key) {
+                tasks.push(task_id);
+            }
+        }
+
+        Ok(tasks)
     }
 
     /// Get the node assigned to a given task
     pub fn get_task_assignment(
         &self,
         task_id: &TaskIdentifier,
-    ) -> Result<Option<WrappedPeerId>, StorageError> {
+    ) -> Result<Option<PeerIdValue>, StorageError> {
         let key = task_assignment_key(task_id);
-        let value = self.inner().read(TASK_ASSIGNMENT_TABLE, &key)?;
-
-        Ok(value)
+        self.inner().read::<_, WrappedPeerId>(TASK_ASSIGNMENT_TABLE, &key)
     }
 }
 
@@ -50,29 +100,37 @@ impl<T: TransactionKind> StateTxn<'_, T> {
 // -----------
 
 impl StateTxn<'_, RW> {
-    /// Push a task to the list of assigned tasks for a node
+    /// Add a task to the list of assigned tasks for a node
+    ///
+    /// This is an O(1) operation - just writes a single key for the
+    /// peer-to-task index plus the task-to-peer lookup key.
     pub fn add_assigned_task(
         &self,
         peer_id: &WrappedPeerId,
         task_id: &TaskIdentifier,
     ) -> Result<(), StorageError> {
-        let key = node_tasks_key(peer_id);
-        self.add_to_set(TASK_ASSIGNMENT_TABLE, &key, task_id)?;
+        // Write the peer-to-task index key
+        let key = assigned_task_key(peer_id, task_id);
+        self.inner().write(TASK_ASSIGNMENT_TABLE, &key, &())?;
 
-        // Set the inverse assignment
-        self.assign_task_to_node(task_id, peer_id)
+        // Write the task-to-peer lookup key
+        self.write_task_assignment(task_id, peer_id)
     }
 
     /// Remove a task from the list of assigned tasks for a node
+    ///
+    /// This is an O(1) operation - just deletes a single key for the
+    /// peer-to-task index plus the task-to-peer lookup key.
     pub fn remove_assigned_task(
         &self,
         peer_id: &WrappedPeerId,
         task_id: &TaskIdentifier,
     ) -> Result<(), StorageError> {
-        let key = node_tasks_key(peer_id);
-        self.remove_from_set(TASK_ASSIGNMENT_TABLE, &key, task_id)?;
+        // Delete the peer-to-task index key
+        let key = assigned_task_key(peer_id, task_id);
+        self.inner().delete(TASK_ASSIGNMENT_TABLE, &key)?;
 
-        // Delete the assignment
+        // Delete the task-to-peer lookup key
         let key = task_assignment_key(task_id);
         self.inner().delete(TASK_ASSIGNMENT_TABLE, &key)?;
         Ok(())
@@ -86,57 +144,44 @@ impl StateTxn<'_, RW> {
         from: &WrappedPeerId,
         to: &WrappedPeerId,
     ) -> Result<Vec<TaskIdentifier>, StorageError> {
-        // Reassign the tasks
-        let new_tasks = self.get_assigned_tasks(from)?;
-        let mut curr_tasks = self.get_assigned_tasks(to)?;
+        // Get all tasks from the source peer
+        let tasks_to_reassign = self.get_assigned_tasks(from)?;
 
-        curr_tasks.extend(new_tasks.clone());
-        self.write_assigned_tasks(to, &curr_tasks)?;
-        self.write_assigned_tasks(from, &vec![])?;
+        // For each task, update both index keys
+        for task_id in tasks_to_reassign.iter() {
+            // Delete old peer-to-task key
+            let old_key = assigned_task_key(from, task_id);
+            self.inner().delete(TASK_ASSIGNMENT_TABLE, &old_key)?;
 
-        // Update the inverse mapping
-        for task_id in new_tasks.iter() {
-            self.assign_task_to_node(task_id, to)?;
+            // Write new peer-to-task key
+            let new_key = assigned_task_key(to, task_id);
+            self.inner().write(TASK_ASSIGNMENT_TABLE, &new_key, &())?;
+
+            // Update the task-to-peer lookup
+            self.write_task_assignment(task_id, to)?;
         }
 
-        Ok(new_tasks)
+        Ok(tasks_to_reassign)
     }
 
     // --- Helpers --- //
 
-    /// Add an assignment from a task id to a node
-    fn assign_task_to_node(
+    /// Write the task-to-peer assignment lookup key
+    fn write_task_assignment(
         &self,
         task_id: &TaskIdentifier,
         peer_id: &WrappedPeerId,
     ) -> Result<(), StorageError> {
         let key = task_assignment_key(task_id);
-        self.inner().write(TASK_ASSIGNMENT_TABLE, &key, peer_id)?;
-        Ok(())
-    }
-
-    /// Write the set of assigned tasks for a node
-    fn write_assigned_tasks(
-        &self,
-        peer_id: &WrappedPeerId,
-        value: &Vec<TaskIdentifier>,
-    ) -> Result<(), StorageError> {
-        let key = node_tasks_key(peer_id);
-        // Delete the key if empty
-        if value.is_empty() {
-            self.inner().delete(TASK_ASSIGNMENT_TABLE, &key)?;
-        } else {
-            self.write_set(TASK_ASSIGNMENT_TABLE, &key, value)?;
-        }
-
-        Ok(())
+        self.inner().write(TASK_ASSIGNMENT_TABLE, &key, peer_id)
     }
 }
 
 #[cfg(test)]
 mod test {
-    use common::types::{gossip::mocks::mock_peer, tasks::TaskIdentifier};
     use itertools::Itertools;
+    use types_gossip::mocks::mock_peer;
+    use types_tasks::TaskIdentifier;
 
     use crate::test_helpers::mock_db;
 
@@ -144,7 +189,7 @@ mod test {
     #[test]
     fn test_assignments() {
         let db = mock_db();
-        let peer_id = mock_peer().get_peer_id();
+        let peer_id = mock_peer().peer_id;
         let task_id = TaskIdentifier::new_v4();
         let tx = db.new_write_tx().unwrap();
 
@@ -157,9 +202,9 @@ mod test {
 
         // Check the assigned tasks => should be empty
         let assigned_tasks = tx.get_assigned_tasks(&peer_id).unwrap();
-        let task_assignment = tx.get_task_assignment(&task_id).unwrap();
+        let task_assignment = tx.get_task_assignment(&task_id).unwrap().unwrap();
         assert_eq!(assigned_tasks, vec![task_id]);
-        assert_eq!(task_assignment, Some(peer_id));
+        assert_eq!(*task_assignment, peer_id);
 
         // Remove the task from the assigned tasks
         tx.remove_assigned_task(&peer_id, &task_id).unwrap();
@@ -168,7 +213,7 @@ mod test {
         let assigned_tasks = tx.get_assigned_tasks(&peer_id).unwrap();
         let task_assignment = tx.get_task_assignment(&task_id).unwrap();
         assert!(assigned_tasks.is_empty());
-        assert_eq!(task_assignment, None);
+        assert!(task_assignment.is_none());
     }
 
     /// Tests reassigning an empty set of tasks
@@ -177,8 +222,8 @@ mod test {
         let db = mock_db();
         let tx = db.new_write_tx().unwrap();
 
-        let from = mock_peer().get_peer_id();
-        let to = mock_peer().get_peer_id();
+        let from = mock_peer().peer_id;
+        let to = mock_peer().peer_id;
         let task_id = TaskIdentifier::new_v4();
 
         // Add a task to the receiver
@@ -188,20 +233,22 @@ mod test {
         // Check that the already existing task is the only task assigned to the
         // receiver
         let assigned_tasks = tx.get_assigned_tasks(&to).unwrap();
-        let task_assignment = tx.get_task_assignment(&task_id).unwrap();
+        let task_assignment = tx.get_task_assignment(&task_id).unwrap().unwrap();
         assert_eq!(assigned_tasks, vec![task_id]);
-        assert_eq!(task_assignment, Some(to));
+        assert_eq!(*task_assignment, to);
     }
 
     /// Tests reassigning tasks from a non-empty set
     #[test]
     fn test_reassign_non_empty() {
+        use std::collections::HashSet;
+
         const N: usize = 10;
         let db = mock_db();
         let tx = db.new_write_tx().unwrap();
 
-        let from = mock_peer().get_peer_id();
-        let to = mock_peer().get_peer_id();
+        let from = mock_peer().peer_id;
+        let to = mock_peer().peer_id;
         let from_tasks = (0..N).map(|_| TaskIdentifier::new_v4()).collect_vec();
         let to_tasks = (0..N).map(|_| TaskIdentifier::new_v4()).collect_vec();
 
@@ -214,16 +261,18 @@ mod test {
         // Reassign the tasks
         tx.reassign_tasks(&from, &to).unwrap();
 
-        let expected_tasks = to_tasks.iter().chain(from_tasks.iter()).cloned().collect_vec();
-        let assigned_tasks = tx.get_assigned_tasks(&to).unwrap();
+        // Compare as sets since cursor returns in key order, not insertion order
+        let expected_tasks: HashSet<_> =
+            to_tasks.iter().chain(from_tasks.iter()).cloned().collect();
+        let assigned_tasks: HashSet<_> = tx.get_assigned_tasks(&to).unwrap().into_iter().collect();
         assert_eq!(assigned_tasks, expected_tasks);
 
         // Check that the tasks are reassigned
         for i in 0..N {
-            let from_task_assignment = tx.get_task_assignment(&from_tasks[i]).unwrap();
-            let to_task_assignment = tx.get_task_assignment(&to_tasks[i]).unwrap();
-            assert_eq!(from_task_assignment, Some(to));
-            assert_eq!(to_task_assignment, Some(to));
+            let from_task_assignment = tx.get_task_assignment(&from_tasks[i]).unwrap().unwrap();
+            let to_task_assignment = tx.get_task_assignment(&to_tasks[i]).unwrap().unwrap();
+            assert_eq!(*from_task_assignment, to);
+            assert_eq!(*to_task_assignment, to);
         }
     }
 }
